script3
 1:
The C library function 
char *fgets(char *str, int n, FILE *stream) reads a line from the specified stream and stores it into the string pointed to by str. It stops when either (n-1) characters are read, the newline character is read, or the end-of-file is reached, whichever comes first.
2:
In the C Programming Language,
 the strstr function searches within the string pointed to by s1 for the string pointed to by s2. It returns a pointer to the first occurrence in s1 of s2.
3:
In the C Programming Language,
 the strcpy function copies the string pointed to by s2 into the object pointed to by s1. It returns a pointer to the destination.
4:
The C library function int fputs(const char *str, FILE *stream) writes a string to the specified stream up to but not including the null character.