Script 6
1:
The #if directive, with the #elif, #else, and #endif directives, controls compilation of portions of a source file. If the expression you write (after the #else) has a nonzero value, the line group immediately following the #else directive is retained in the translation unit.
2:
The C library function int fclose(FILE *stream) closes the stream. All buffers are flushed.